requirements.txt
# Updated requirements.txt:
torch==2.5.1  # Or another stable version, depending on your needs.
fastai==2.7.18
timm==1.0.12
numpy==1.26.4
matplotlib==3.9.2
tqdm==4.66.5
hydra-core>=1.3.2
pytorch-lightning==2.3.0
dvc[gs]
fastapi==0.115.6
httpx==0.28.1
functions-framework==3.*
uvicorn==0.34
Pillow==10.4.0
google-cloud-storage==2.8.0  # Updated Google Cloud package
